# artifacts/open_api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Chat APIs
  version: 0.215.37
  description: API for the Simple Chat, handling chats, conversations, documents, user settings, groups, feedback, prompts, models, safety, and admin functions.
servers:
  - url: /api
    description: API Root
...
paths:
...
  /prompts:
    get:
      tags:
        - Prompts
      summary: Get user's saved prompts (paginated)
      description: Retrieves a paginated list of prompts saved by the authenticated user. Requires `enable_user_workspace` setting to be true. Filtering via query params may be supported by `list_prompts`.
      security:
        - bearerAuth: []
      parameters:
        # Add pagination/filtering params based on list_prompts implementation
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: The page number.
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
          description: Items per page.
        - name: search
          in: query
          schema:
            type: string
          description: Optional search term for prompt name.
      responses:
        '200':
          description: User prompts retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPrompt'
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_count:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database query failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Prompts
      summary: Create a new user prompt
      description: Creates and saves a new personal prompt for the authenticated user. Requires `enable_user_workspace` setting to be true.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - content
              properties:
                name:
                  type: string
                  description: The name/title of the prompt. Cannot be empty.
                content:
                  type: string
                  description: The text content of the prompt. Cannot be empty.
      responses:
        '201': # Code returns 201
          description: User prompt created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPrompt' # Returns the created prompt
        '400':
          description: Bad Request (Missing or invalid name or content)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database create failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /prompts/{prompt_id}:
    get:
      tags:
        - Prompts
      summary: Get a specific user prompt
      description: Retrieves the details of a single prompt saved by the authenticated user. Requires `enable_user_workspace` setting to be true. Checks ownership.
      security:
        - bearerAuth: []
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user prompt to retrieve.
      responses:
        '200':
          description: User prompt retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPrompt'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Prompt not found or access denied (not owned by user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database read failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Prompts
      summary: Update a user prompt
      description: Updates the name and/or content of a specific prompt saved by the authenticated user. Requires `enable_user_workspace` setting to be true. Checks ownership.
      security:
        - bearerAuth: []
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user prompt to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1 # Must provide at least name or content
              properties:
                name:
                  type: string
                  description: The new name for the prompt (must be non-empty if provided).
                content:
                  type: string
                  description: The new content for the prompt.
      responses:
        '200':
          description: User prompt updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPrompt' # Returns the updated prompt
        '400':
          description: Bad Request (Invalid or missing name/content, or no fields provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Prompt not found or access denied (not owned by user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database update failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Prompts
      summary: Delete a user prompt
      description: Deletes a specific prompt saved by the authenticated user. Requires `enable_user_workspace` setting to be true. Checks ownership.
      security:
        - bearerAuth: []
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user prompt to delete.
      responses:
        '200':
          description: User prompt deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Prompt deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_user_workspace` disabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Prompt not found or access denied (not owned by user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database delete failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --- Safety Routes (route_backend_safety.py) ---
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
...
    UserPrompt:
      type: object
      description: A personal prompt saved by a user.
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          description: ID of the user who owns this prompt.
        name:
          type: string
          description: Name or title of the prompt.
        content:
          type: string
          description: The text content of the prompt.
        type:
          type: string
          example: "user_prompt"
          description: Internal type identifier.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the prompt was created.
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the prompt was last updated.
...
  securitySchemes:
    bearerAuth: # Defines how API clients authenticate
      type: http
      scheme: bearer
      bearerFormat: JWT # Indicates the type of token expected
      description: "Azure AD issued JWT Bearer token required for authenticated endpoints. Include in the request header as 'Authorization: Bearer <YOUR_TOKEN>'."